requirements.txt
peewee==3.13.2
pyyaml==5.3.1
pymysql==0.9.3